+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 
-- `dp publish` will push generated sources to external git repo
+## [0.12.0] - 2022-01-31
+
+-   `dp publish` will push generated sources to external git repo
 
 ## [0.11.0] - 2022-01-18
 
@@ -146,7 +148,9 @@
 -   Draft of `dp init`, `dp create`, `dp template new`, `dp template list` and `dp dbt`
 -   Draft of `dp compile` and `dp deploy`
 
-[Unreleased]: https://github.com/getindata/data-pipelines-cli/compare/0.11.0...HEAD
+[Unreleased]: https://github.com/getindata/data-pipelines-cli/compare/0.12.0...HEAD
+
+[0.12.0]: https://github.com/getindata/data-pipelines-cli/compare/0.11.0...0.12.0
 
 [0.11.0]: https://github.com/getindata/data-pipelines-cli/compare/0.10.0...0.11.0